feat: Add WLED device control and management utilities
- Implemented device operations including heartbeat check, power control, brightness adjustment, and state retrieval in wledDevice.ts.
- Created gradient generation utilities for playlists and presets in wledGradients.ts.
- Developed preset and playlist management functions in wledPresets.ts, including fetching, creating, and deleting presets and playlists.
- Added timer management functions for scheduling in wledTimer.ts, including fetching, saving, and resetting timer settings.
- Introduced utility functions for API interactions and data parsing in wledUtils.ts.
